python-services/email-service/requirements.txt
fastapi
uvicorn
python-dotenv
smtplib
email-validator
 